Add files via upload
This code creates a new class called AssignmentOne. And in the "AssignmentOne" class, create the main method. It is a Java program that simulates the health service appointment management system. It allows patients to make appointments with different types of health professionals, such as general practitioners and doctors in other specialized areas, and to view and cancel appointments.