Add files via upload
The purpose of this class is to define a basic information class for a health professional that has three main attributes: id, name, and specialization. It provides the following features:
Constructor: Allows you to create an instance of the HealthProfessional class and initialize its ID, name, and specialty area.
The printDetails() method: prints the health professional's details.
The HealthProfessional class defines a basic data structure for a health professional and provides initialization and printing capabilities. In practical applications, it can be used as a base class for other health professional classes to build more complex systems through inheritance and extension.