Add files via upload
This code defines a Java class called Specialist, which inherits from the HealthProfessional class, and adds a new attribute to it. The purpose of this class is to extend the HealthProfessional class. Denotes a specific type of health professional - specialist. Compared to the base class, it adds a new instance variable, through the printDetails() method, the program can output the details of the expert, including the expert's area of expertise.